more precedence climbing and switch syntax change
multiple switch cases are now separated by , instead of |
because | is bitwise or, and the precedence effect is actually really strange.

the only reason this seems fine in rust is because their match condition is
actually a separate precedence list?

that is actually pretty stupid, but whatever.

also i finally fixed nested struct access by putting memberaccess inside
the precedence climber.