change fntype to closures
this is a rather big breaking change.
In order to support mapping zz api to dynamic languages, callbacks need
to carry interpreter context. there's no clean way to automatically
generate bindings from a C api, so zetz now has a
high level construct instead, which is the closure.

for MOST code it should simply be a matter of replacing 'fntype' with
'closure', but the C behaviour is different and unsafe code will break