constructor uses new arg by tag rather than position
this is a breaking change. code that previously incorrectly ommited the
"new" tag on constructor functions won't compile anymore until the new
tag is added.

the goal is to allow mixing self calls with constructors like this:

  new a = a::make();
  new b = a.to_string();

that requires expand to know the type of a, which is on TODO anyway