BREAKING fix macro string parsing
string is a vec<u8> due to different bytestring interpretations,
so ast has to parse this as vec.

While we're at it, this is a breaking change to move ast to a
dynamically allocated tree, preparing for struct macros (derive).

refer to tests/mustpass/macro_expr for how the api changed