implement attached type expressions:
this introduces smt expressions that can be attached permanently to a type.
The primary use case is generic containers.

see #121

attaching expressions:

    struct A {}
    theory foo(A a) bool;

    fn main() {
        A[foo(self) == true] a;
    }

which is equivalent to

    A a;
    static_attest(foo(a) == true);

except the expression is carried for the lifetime of a instead of for
the lifetime of the value of a, meaning it is a property of the type A
of a rather than a property of a